Add files via upload
An E-commerce company is facing a lot of competition in the current market and it has become a challenge to retain the existing customers in the current situation. 
Hence, the company wants to develop a model through which they can do churn prediction of the accounts and provide segmented offers to the potential churners.